chore: created album entity and moved ui
