+++ src/entity/albums/constants.ts
@@ -1,0 +1,18 @@
+import type { Album } from "@/types";
+
+export const ALBUMS: Album[] = [
+  { id: 1, title: "Ядрена вошь", content: "новый альбом", backgroundImage: "yadrena-vosh.jpg" },
+  { id: 2, title: "Куда преш", content: "новый альбом", backgroundImage: "kuda-presh.jpg" },
+  {
+    id: 3,
+    title: "Тапки в тряпки",
+    content: "новый альбом",
+    backgroundImage: "tapki.jpg",
+  },
+  {
+    id: 4,
+    title: "Крути меня мошпит",
+    content: "новый альбом",
+    backgroundImage: "moshpit.jpg",
+  },
+];
