+++ src/entity/albums/index.ts
@@ -1,3 +1,3 @@
+export * from "./constants";
 export * from "./view";
 export * from "./model";
-export * from "./constants";
