+++ src/entity/albums/model.ts
@@ -1,0 +1,1 @@
+export {};
