+++ src/entity/playlists/constants.ts
@@ -1,0 +1,1 @@
+export {};
